Split off constants
This is in preparation for constant expansion and handling
new settign Settings`$PreferredBackendMethod

blacken some buffers too.
