+++ mathics/builtin/__init__.py
@@ -3,30 +3,108 @@
 
 
 from mathics.builtin import (
-    algebra, arithmetic, assignment, attributes, calculus, combinatorial, compilation,
-    comparison, control, datentime, diffeqns, evaluation, exptrig, functional,
-    graphics, graphics3d,
-    image, inout, integer, iohooks, linalg, lists, logic,
-    manipulate, quantities, numbertheory, numeric, options, patterns,
-    plot, physchemdata, randomnumbers, recurrence, specialfunctions, scoping,
-    strings, structure, system, tensors, xmlformat, optimization)
+    algebra,
+    arithmetic,
+    assignment,
+    attributes,
+    calculus,
+    combinatorial,
+    compilation,
+    comparison,
+    constants,
+    control,
+    datentime,
+    diffeqns,
+    evaluation,
+    exptrig,
+    functional,
+    graphics,
+    graphics3d,
+    image,
+    inout,
+    integer,
+    iohooks,
+    linalg,
+    lists,
+    logic,
+    manipulate,
+    quantities,
+    numbertheory,
+    numeric,
+    options,
+    patterns,
+    plot,
+    physchemdata,
+    randomnumbers,
+    recurrence,
+    specialfunctions,
+    scoping,
+    strings,
+    structure,
+    system,
+    tensors,
+    xmlformat,
+    optimization,
+)
 
 from mathics.builtin.base import (
-    Builtin, SympyObject, BoxConstruct, Operator, PatternObject)
+    Builtin,
+    SympyObject,
+    BoxConstruct,
+    Operator,
+    PatternObject,
+)
 
 from mathics.settings import ENABLE_FILES_MODULE
 
 modules = [
-    algebra, arithmetic, assignment, attributes, calculus, combinatorial, compilation,
-    comparison, control, datentime, diffeqns, evaluation, exptrig, functional,
-    graphics, graphics3d,
-    image, inout, integer, iohooks, linalg, lists, logic,
-    manipulate, quantities, numbertheory, numeric, options, patterns,
-    plot, physchemdata, randomnumbers, recurrence, specialfunctions, scoping,
-    strings, structure, system, tensors, xmlformat, optimization]
+    algebra,
+    arithmetic,
+    assignment,
+    attributes,
+    calculus,
+    combinatorial,
+    compilation,
+    comparison,
+    constants,
+    control,
+    datentime,
+    diffeqns,
+    evaluation,
+    exptrig,
+    functional,
+    graphics,
+    graphics3d,
+    image,
+    inout,
+    integer,
+    iohooks,
+    linalg,
+    lists,
+    logic,
+    manipulate,
+    quantities,
+    numbertheory,
+    numeric,
+    options,
+    patterns,
+    plot,
+    physchemdata,
+    randomnumbers,
+    recurrence,
+    specialfunctions,
+    scoping,
+    strings,
+    structure,
+    system,
+    tensors,
+    xmlformat,
+    optimization,
+]
 
 if ENABLE_FILES_MODULE:
     from mathics.builtin import files, importexport
+
     modules += [files, importexport]
 
 builtins = []
@@ -36,7 +114,7 @@ builtins_by_module = {}
 def is_builtin(var):
     if var == Builtin:
         return True
-    if hasattr(var, '__bases__'):
+    if hasattr(var, "__bases__"):
         return any(is_builtin(base) for base in var.__bases__)
     return False
 
@@ -46,11 +124,14 @@ for module in modules:
     vars = dir(module)
     for name in vars:
         var = getattr(module, name)
-        if (hasattr(var, '__module__') and
-            var.__module__.startswith('mathics.builtin.') and
-            var.__module__ != 'mathics.builtin.base' and
-            is_builtin(var) and not name.startswith('_') and
-            var.__module__ == module.__name__):     # nopep8
+        if (
+            hasattr(var, "__module__")
+            and var.__module__.startswith("mathics.builtin.")
+            and var.__module__ != "mathics.builtin.base"
+            and is_builtin(var)
+            and not name.startswith("_")
+            and var.__module__ == module.__name__
+        ):  # nopep8
 
             instance = var(expression=False)
 
@@ -61,8 +142,8 @@ for module in modules:
 
 # builtins = dict(builtins)
 
-mathics_to_sympy = {} # here we have: name -> sympy object
-mathics_to_python = {} # here we have: name -> string
+mathics_to_sympy = {}  # here we have: name -> sympy object
+mathics_to_python = {}  # here we have: name -> string
 sympy_to_mathics = {}
 
 box_constructs = {}
@@ -101,22 +182,22 @@ def get_module_doc(module):
         doc = doc.strip()
     if doc:
         title = doc.splitlines()[0]
-        text = '\n'.join(doc.splitlines()[1:])
+        text = "\n".join(doc.splitlines()[1:])
     else:
         title = module.__name__
-        for prefix in ('mathics.builtin.', 'mathics.optional.'):
+        for prefix in ("mathics.builtin.", "mathics.optional."):
             if title.startswith(prefix):
-                title = title[len(prefix):]
+                title = title[len(prefix) :]
         title = title.capitalize()
-        text = ''
+        text = ""
     return title, text
 
 
 def contribute(definitions):
     # let MakeBoxes contribute first
-    builtins['System`MakeBoxes'].contribute(definitions)
+    builtins["System`MakeBoxes"].contribute(definitions)
     for name, item in builtins.items():
-        if name != 'System`MakeBoxes':
+        if name != "System`MakeBoxes":
             item.contribute(definitions)
 
     from mathics.core.expression import ensure_context
