+++ mathics/builtin/base.py
@@ -28,22 +28,6 @@ from mathics.core.expression import (
 )
 from mathics.core.numbers import get_precision, PrecisionValueError
 
-def mp_fn(fn, d=None):
-    if d is None:
-        return getattr(mpmath, fn)()
-    else:
-        mpmath.mp.dps = int_d = int(d)
-        return getattr(mpmath, fn)(prec=int_d)
-
-def mp_convert_constant(obj, **kwargs):
-    if isinstance(obj, mpmath.ctx_mp_python._constant):
-        prec = kwargs.get("prec", None)
-        if prec is not None:
-            return sympy.Float(obj(prec=prec))
-        return sympy.Float(obj)
-    return obj
-
-
 def get_option(options, name, evaluation, pop=False, evaluate=True):
     # we do not care whether an option X is given as System`X,
     # Global`X, or with any prefix from $ContextPath for that
@@ -444,6 +428,26 @@ class Predefined(Builtin):
         return functions
 
 
+class SympyObject(Builtin):
+    sympy_name: typing.Optional[str] = None
+
+    mathics_to_sympy = {}
+
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        if self.sympy_name is None:
+            self.sympy_name = strip_context(self.get_name()).lower()
+        self.mathics_to_sympy[self.__class__.__name__] = self.sympy_name
+
+    def is_constant(self) -> bool:
+        return False
+
+    def get_sympy_names(self) -> typing.List[str]:
+        if self.sympy_name:
+            return [self.sympy_name]
+        return []
+
+
 class UnaryOperator(Operator):
     def __init__(self, format_function, *args, **kwargs):
         super().__init__(*args, **kwargs)
@@ -528,144 +532,6 @@ class Test(Builtin):
             return Symbol("True")
         else:
             return Symbol("False")
-
-
-class MPMathConstant(Predefined):
-    """Representation of a constant in mpmath, e.g. Pi, E, I, etc."""
-
-    attributes = ("Constant", "ReadProtected")
-
-    # Subclasses should define this.
-    mpmath_name = None
-
-    mathics_to_mpmath = {}
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if self.mpmath_name is None:
-            self.mpmath_name = strip_context(self.get_name()).lower()
-        self.mathics_to_mpmath[self.__class__.__name__] = self.mpmath_name
-
-    def get_constant(self, precision, evaluation, preference="mpmath"):
-        try:
-            d = get_precision(precision, evaluation)
-        except PrecisionValueError:
-            return
-
-        mpmath_name = self.mpmath_name
-        if d is None:
-            return MachineReal(mp_fn(mpmath_name))
-        elif preference == "mpmath":
-            result = mp_fn(mpmath_name, d)
-        else:
-            sympy_fn = self.to_sympy()
-            try:
-                result = sympy_fn.n(d)
-            except:
-                from trepan.api import debug; debug()
-                pass
-        return PrecisionReal(result)
-
-
-    def is_constant(self) -> bool:
-        # free Symbol will be converted to corresponding SymPy symbol
-        return True
-
-    def to_mpmath(self, args):
-        if self.mpmath_name is None or len(args) != self.nargs:
-            return None
-        return getattr(mpmath, self.mpmath_name)
-
-
-class NumpyConstant(Predefined):
-    """Representation of a constant in numpy, e.g. Pi, E, etc."""
-
-    attributes = ("Constant", "ReadProtected")
-
-    # Subclasses should define this.
-    numpy_name = None
-
-    mathics_to_numpy = {}
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if self.numpy_name is None:
-            self.numpy_name = strip_context(self.get_name()).lower()
-        self.mathics_to_numpy[self.__class__.__name__] = self.numpy_name
-
-    def is_constant(self) -> bool:
-        # free Symbol will be converted to corresponding SymPy symbol
-        return True
-
-    def to_numpy(self, args):
-        if self.numpy_name is None or len(args) != self.nargs:
-            return None
-        return getattr(numpy, self.numpy_name)
-
-
-class SympyObject(Builtin):
-    sympy_name: typing.Optional[str] = None
-
-    mathics_to_sympy = {}
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        if self.sympy_name is None:
-            self.sympy_name = strip_context(self.get_name()).lower()
-        self.mathics_to_sympy[self.__class__.__name__] = self.sympy_name
-
-    def is_constant(self) -> bool:
-        return False
-
-    def get_sympy_names(self) -> typing.List[str]:
-        if self.sympy_name:
-            return [self.sympy_name]
-        return []
-
-
-class SympyConstant(SympyObject, Predefined):
-    """Representation of a constant in Sympy, e.g. Pi, E, I, Catalan, etc."""
-
-    attributes = ("Constant", "ReadProtected")
-
-    # Subclasses should define this.
-    sympy_name = None
-
-    def get_constant(self, precision, evaluation, preference="sympy"):
-        try:
-            d = get_precision(precision, evaluation)
-        except PrecisionValueError:
-            return
-
-        sympy_fn = self.to_sympy()
-        if d is None:
-            if hasattr(self, "mpmath_name"):
-                # Prefer mpmath when precision is not given.
-                result = mp_fn(self.mpmath_name)
-            else:
-                result = sympy_fn()
-            return MachineReal(result)
-
-        if preference == "sympy":
-            result = sympy_fn_n(d)
-        elif hasattr(self, "mpmath_name"):
-            result = mp_fn(self.mpmath_name, d)
-
-        return PrecisionReal(result)
-
-    def is_constant(self) -> bool:
-        # free Symbol will be converted to corresponding SymPy symbol
-        return True
-
-    def to_sympy(self, expr=None, **kwargs):
-        if expr is None or expr.is_atom():
-            result = getattr(sympy, self.sympy_name)
-            if kwargs.get("evaluate", False):
-                result = mp_convert_constant(result, **kwargs)
-            return result
-        else:
-            # there is no "native" SymPy expression for e.g. E[x]
-            return None
 
 
 class SympyFunction(SympyObject):
