+++ mathics/version.py
@@ -5,4 +5,4 @@
 # This file is suitable for sourcing inside POSIX shell as
 # well as importing into Python. That's why there is no
 # space around "=" below.
-__version__="2.0.0dev"  # noqa
+__version__="2.0.0.dev0"  # noqa
