+++ test/test_combinatorica.py
@@ -49,6 +49,7 @@ def test_permutations_1_1():
             " {3, 1, 2, 4}, {3, 1, 4, 2}, {3, 2, 1, 4}, {3, 2, 4, 1}, "
             " {3, 4, 1, 2}, {3, 4, 2, 1}, {4, 1, 2, 3}, {4, 1, 3, 2}, "
             " {4, 2, 1, 3}, {4, 2, 3, 1}, {4, 3, 1, 2}, {4, 3, 2, 1}} ",
+
             "slower method for computing permutations in lex order, 1.1.2, Page 6",
         ),
         (
